Refactoring: remove fields with Array type
