+++ Autopark.API/wwwroot/js/components/Vehicle/AddVehicleForm.js
@@ -1,0 +1,9 @@
+class AddVehicleFormComponent extends HTMLElement {
+    async connectedCallback() {
+        //...
+    }
+}
+
+export const registerVehiclesTableComponent = () => {
+    customElements.define('x-add-vehicle-form', AddVehicleFormComponent)
+}
